add pipeline options, move ci.py into its own module
